backend/requirements.txt
grpcio>=1.59.0
grpcio-tools>=1.59.0
sqlalchemy>=2.0.0
protobuf>=4.25.0
psycopg[binary]>=3.1
python-dotenv>=1.0.1
flask>=2.3.0
flask-cors>=4.0.0
pydantic>=2.0.0
email-validator>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
factory-boy>=3.2.0
faker>=18.0.0 